+++ talpid-core/src/linux/network_manager.rs
@@ -83,6 +83,9 @@ pub enum Error {
     #[error(display = "NetworkManager not detected")]
     NetworkManagerNotDetected,
 
+    #[error(display = "NetworkManager is using dnsmasq to manage DNS")]
+    UsingDnsmasq,
+
     #[error(display = "NetworkManager is too old: {}", 0)]
     TooOldNetworkManager(String),
 
@@ -491,6 +494,8 @@ impl NetworkManager {
             .map_err(Error::Dbus)?;
 
         match dns_mode.as_ref() {
+            // NM can't setup config for multiple interfaces with dnsmasq
+            "dnsmasq" => return Err(Error::UsingDnsmasq),
             // If NetworkManager isn't managing DNS for us, it's useless.
             "none" => return Err(Error::NetworkManagerNotManagingDns),
             _ => (),